added lines for checking tokens in posts endpoints

+++ DrGym/src/main/java/com/drgym/drgym/controller/PostController.java
@@ -5,6 +5,8 @@ import com.drgym.drgym.service.ExerciseService;
 import com.drgym.drgym.service.PostService;
 import com.drgym.drgym.service.PostReactionService;
 import com.drgym.drgym.service.WorkoutService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -27,16 +29,21 @@ public class PostController {
     private WorkoutService workoutService;
 
     @Autowired
-    private ExerciseService exerciseService;
+    private UserController userController;
 
     @GetMapping("/{postId}")
-    public ResponseEntity<?> getPost(@PathVariable Long postId) {
+    public ResponseEntity<?> getPost(@PathVariable Long postId, HttpServletRequest request) {
         Optional<Post> postOptional = postService.findPostById(postId);
         if (postOptional.isEmpty()) {
             return ResponseEntity.notFound().build();
         }
 
         Post post = postOptional.get();
+        String username = post.getUsername();
+        if (!userController.tokenOwnerOrFriend(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+
         Workout workout = post.getTraining();
         if (workout != null) {
             List<Activity> activities = workoutService.findActivitiesByWorkoutId(workout.getId());
@@ -47,26 +54,57 @@ public class PostController {
     }
 
     @GetMapping("/user/{username}")
-    public ResponseEntity<?> getUserPosts(@PathVariable String username) {
+    public ResponseEntity<?> getUserPosts(@PathVariable String username, HttpServletRequest request) {
+        if (!userController.tokenOwnerOrFriend(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+
         List<Post> posts = postService.findPostsByUsername(username);
         if (posts.isEmpty()) {
             return ResponseEntity.ok("[]");
         }
+
+        posts.forEach(post -> {
+            Workout workout = post.getTraining();
+            if (workout != null) {
+                List<Activity> activities = workoutService.findActivitiesByWorkoutId(workout.getId());
+                workout.setActivities(activities);
+            }
+        });
+
         return ResponseEntity.ok(posts);
     }
 
     @PostMapping("/create")
-    public ResponseEntity<?> createPost(@RequestBody PostCreateRequest postRequest) {
+    public ResponseEntity<?> createPost(@RequestBody PostCreateRequest postRequest, HttpServletRequest request) {
+        if (!userController.tokenOwner(postRequest.getUsername(), request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
         return postService.createPost(postRequest);
     }
 
     @PostMapping("/create_with_workout")
-    public ResponseEntity<?> createPost(@RequestBody PostCreateRequestWorkout postRequest) {
+    public ResponseEntity<?> createPost(@RequestBody PostCreateRequestWorkout postRequest, HttpServletRequest request) {
+        if (!userController.tokenOwner(postRequest.getUsername(), request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
         return postService.createPostWorkout(postRequest);
     }
 
     @GetMapping("/{postId}/reactions")
-    public ResponseEntity<?> getReactionsForPost(@PathVariable Long postId) {
+    public ResponseEntity<?> getReactionsForPost(@PathVariable Long postId, HttpServletRequest request) {
+        Optional<Post> postOptional = postService.findPostById(postId);
+        if (postOptional.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        Post post = postOptional.get();
+        String username = post.getUsername();
+
+        if (!userController.tokenOwnerOrFriend(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+
         List<PostReaction> reactions = postReactionService.findByPostId(postId);
         if (reactions.isEmpty()) {
             return ResponseEntity.notFound().build();
@@ -75,14 +113,21 @@ public class PostController {
     }
 
     @PostMapping("/{postId}/reactions")
-    public ResponseEntity<?> addReactionToPost(@PathVariable Long postId, @RequestParam String username) {
+    public ResponseEntity<?> addReactionToPost(@PathVariable Long postId, @RequestParam String username, HttpServletRequest request) {
+        if (!userController.tokenOwnerOrFriend(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
         PostReaction reaction = new PostReaction(postId, username);
         postReactionService.addReaction(reaction);
         return ResponseEntity.ok("Reaction added successfully.");
     }
 
     @DeleteMapping("/{postId}/reactions")
-    public ResponseEntity<?> deleteReaction(@PathVariable Long postId, @RequestParam String username) {
+    public ResponseEntity<?> deleteReaction(@PathVariable Long postId, @RequestParam String username, HttpServletRequest request) {
+        if (!userController.tokenOwnerOrFriend(username, request)) {
+            return ResponseEntity.status(HttpServletResponse.SC_UNAUTHORIZED).body("Unauthorized");
+        }
+
         postReactionService.removeReactionByUsername(postId, username);
         return ResponseEntity.noContent().build();
     }
